+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Releases the Local Emulator UI, which now starts with `firebase emulators:start`.
 - Fixes bug causing Hosting emulator to serve invalid /\_\_/\* files.
 - Fixes bug in Firestore emulator for evaluating rules in query listening. (#2197)
 - Fixes support in Firestore emulator for `?show_missing` in listing collection.